102.2

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,18 +8,18 @@
 
 int main(void)
 {
-	long long int i1 = 1;
-	long long int i2 = 2;
-	long long int i3;
-	long long int n;
+	unsigned long int i1 = 1;
+	unsigned long int i2 = 2;
+	unsigned long int i3;
+	unsigned long int n;
 
-	printf("%lld, ", i1);
-	printf("%lld, ", i2);
+	printf("%lu, ", i1);
+	printf("%lu, ", i2);
 
 	for (n = 3; n <= 50; n++)
 	{
 		i3 = i1 + i2;
-		printf("%lld, ", i3);
+		printf("%lu, ", i3);
 		i1 = i2;
 		i2 = i3;
 	}
